% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.ggeffects}
\alias{autoplot.ggeffects}
\title{Create a (blank) ggplot from ggeffects}
\usage{
\method{autoplot}{ggeffects}(object, mapping = NULL, fct_rows = NULL, fct_cols = NULL)
}
\arguments{
\item{object}{A \code{ggeffects} object.}

\item{mapping}{Default list of aesthetic mappings to use for plot. If not
specified, default is to map \code{predicted} to \code{y}, the first term to \code{x} and
the second term to \code{group}, \code{color} and \code{fill}. \code{conf.high}/\code{conf.low} are
mapped onto \code{ymax}/\code{ymin}.}

\item{fct_rows, fct_cols}{Specs for the facetting of the plot. If left \code{NULL}
will be taken from the \code{facet} and \code{panel} columns.}
}
\description{
Create a (blank) ggplot from ggeffects
}
