% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_data.R
\name{extract_fit_data}
\alias{extract_fit_data}
\title{Extract data from a \code{ggeffects} object}
\usage{
extract_fit_data(object, residualize = FALSE, collapse.group = NULL)
}
\arguments{
\item{object}{A \code{ggeffects} object.}

\item{residualize}{Return partially residualized data instead of the raw
data?}

\item{collapse.group}{For multilevel models, the name of the random grouping
variable over which data should be aggregated prior to plotting. If \code{TRUE},
the first grouping variable will be used.}
}
\value{
A data frame
}
\description{
Extract the data used to fit the model on which the \code{ggeffects} grid is
based. Data can be residualized (with \code{\link[=residualize_over_grid]{residualize_over_grid()}}) and/or
collapsed across a grouping variable.
}
\seealso{
\code{\link[=layer_fit_data]{layer_fit_data()}}
}
