% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_fit_data.R
\name{layer_fit_data}
\alias{layer_fit_data}
\title{Add data to plot}
\usage{
layer_fit_data(
  mapping = NULL,
  geom = "point",
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  residuals = FALSE,
  collapse.by = NULL,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}} or
\code{\link[ggplot2:aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{geom}{The geometric object to use display the data}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}

\item{residuals}{Return partially residualized data instead of the raw
data?}

\item{collapse.by}{For multilevel models, the name of the random grouping
variable over which data should be aggregated prior to plotting. If \code{TRUE},
the first grouping variable will be used.}
}
\description{
Add the data used to fit the model on which the \code{ggeffects} grid is based.
Data can be residualized (with \code{\link[=residualize_over_grid]{residualize_over_grid()}}) for \emph{partial
residual plots} and/or collapsed across a grouping variable (with
\code{\link[=collapse_by_group]{collapse_by_group()}}).
}
\section{Computed variables}{

\itemize{
\item The 1-to-4 terms (by their name)
\item \code{predicted} - The response / predicted outcome + residuals
}
}

