% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.ggeffects}
\alias{autoplot.ggeffects}
\title{Create a (blank) ggplot from ggeffects}
\usage{
\method{autoplot}{ggeffects}(object, mapping = NULL, rows = waiver(), cols = waiver())
}
\arguments{
\item{object}{A \code{ggeffects} object.}

\item{mapping}{Default list of aesthetic mappings to use for plot. If not
specified, default is to map \code{predicted} to \code{y}, the first term to \code{x} and
the second term to \code{group}, \code{color} and \code{fill}. \code{conf.high}/\code{conf.low} are
mapped onto \code{ymax}/\code{ymin}.}

\item{rows, cols}{Specs for the facetting of the plot. See \code{rows} and
\code{cols} args in \code{facet_grid()} Defaults to \code{facet} and \code{panel} columns.
Disable facetting by setting both to \code{NULL}.}
}
\description{
Create a (blank) ggplot from ggeffects
}
